+++ js-api/src/interfaces/d4.ts
@@ -495,6 +495,30 @@ export interface IScatterPlotSettings {
 
   showYSelector: boolean;
 
+  /// Point lower bound for x axis whiskers.
+  xWhiskerMin: string;
+  xWhiskerMinColumnName: string;
+
+  /// Point upper bound for x axis whiskers.
+  xWhiskerMax: string;
+  xWhiskerMaxColumnName: string;
+
+  /// Point range for x axis whiskers.
+  xWhiskerRange: string;
+  xWhiskerRangeColumnName: string;
+
+  /// Point lower bound for y axis whiskers.
+  yWhiskerMin: string;
+  yWhiskerMinColumnName: string;
+
+  /// Point upper bound for y axis whiskers.
+  yWhiskerMax: string;
+  yWhiskerMaxColumnName: string;
+
+  /// Point range for y axis whiskers.
+  yWhiskerRange: string;
+  yWhiskerRangeColumnName: string;
+
   xAxisLabelOrientation: string;
 
   /// A column to be used for color-coding. Could be numerical or categorical.
@@ -641,6 +665,8 @@ export interface IScatterPlotSettings {
   categoricalColorScheme: Array<number>;
 
   regressionLineColor: number;
+
+  whiskerColor: number;
 
   regressionLineTransparency: number;