+++ packages/ApiSamples/scripts/data-frame/columns-ordering.js
@@ -1,0 +1,4 @@
+
+const demog = grok.data.demo.demog(100);
+demog.columns.setOrder(['height', 'weight', 'age']);
+grok.shell.addTableView(demog);