+++ packages/DiffStudio/src/scripting-tools.ts
@@ -8,7 +8,7 @@
 */
 
 import {CONTROL_TAG, CONTROL_TAG_LEN, DF_NAME, CONTROL_EXPR, LOOP, UPDATE, MAX_LINE_CHART,
-  SOLVER_OPTIONS_RANGES} from './constants';
+  SOLVER_OPTIONS_RANGES, TINY, STEP_RATIO} from './constants';
 
 // Scripting specific constants
 export const CONTROL_SEP = ':';
@@ -975,6 +975,9 @@ function getScriptMainBodyLoopCase(ivp: IVP): string[] {
       res.push(`${SCRIPT.SPACE2}${name} = ${DF_NAME}.get('${name}', ${SCRIPT.LAST_IDX});`);
   });
 
+  // eslint-disable-next-line max-len
+  res.push(`${SCRIPT.SPACE2}${DF_NAME}.set('${ivp.arg.name}', ${SCRIPT.LAST_IDX}, ${SERVICE}${ivp.arg.name}0 - Math.min(${SERVICE}h * ${STEP_RATIO}, ${TINY}));`);
+
   res.push('};');
 
   return res;
@@ -1006,6 +1009,9 @@ function getScriptMainBodyUpdateCase(ivp: IVP): string[] {
     res.push(`${SCRIPT.UPDATE_COM} ${idx + 1}`);
     res.push(`const ${UPDATE.DURATION}${idx + 1} = ${upd.durationFormula};`);
     res.push(`${SCRIPT.LAST_IDX} = ${DF_NAME}.rowCount - 1;`);
+
+    // eslint-disable-next-line max-len
+    res.push(`${DF_NAME}.set('${ivp.arg.name}', ${SCRIPT.LAST_IDX}, ${SERVICE}${ivp.arg.name}1 - Math.min(${SERVICE}h * ${STEP_RATIO}, ${TINY}));`);
 
     dfNames.forEach((name, idx) => {
       if (idx !== 0)