+++ packages/PowerPack/src/widgets/about-widget.ts
@@ -5,14 +5,12 @@ import * as DG from 'datagrok-api/dg';
 
 export class AboutWidget extends DG.Widget {
   caption: string;
-  order: string;
 
   constructor() {
     super(ui.box());
 
     // properties
     this.caption = super.addProperty('caption', DG.TYPE.STRING, 'Why Datagrok?');
-    this.order = super.addProperty('order', DG.TYPE.STRING, '4');
 
     let tabControl = ui.tabControl({
     'PLATFORM' : () => ui.panel([ui.markdown(intro)]),
@@ -65,7 +63,7 @@ export class AboutWidget extends DG.Widget {
   }
 }
 
-var intro =  `Datagrok unlocks the value of the complex data by empowering 
+const intro =  `Datagrok unlocks the value of the complex data by empowering 
 non-technical users to 
 [discover](/help/discover/fair), 
 [cleanse](/help/transform/data-wrangling), 
@@ -79,14 +77,15 @@ is missing, extend the platform with  [plugins and applications](/help/develop/d
 The platform is a lot more than a tool for data analysis - it is an ecosystem designed
 for users, IT, data scientists, chemists, and developers not only to collaborate, but to 
 create value for other types of users as well. Click on the tabs to the left to see how 
-the platform would benefit different users.
-`;
-var introUsers = `Access any data source, anytime. Use the same platform for
+the platform would benefit different users.`;
+
+const introUsers = `Access any data source, anytime. Use the same platform for
 data discovery, self-service data analytics, creating and sharing dashboards.
 Run fit-for-purpose applications developed on top of the platform.
 * Keep your files in order, access from any computer, and share with others
 * Get [context-specific suggestions](/help/discover/data-augmentation) for analyzing data in front of you`;
-var introIT = `Software is eating the world, and IT is a critical competence for all enterprises.
+
+const introIT = `Software is eating the world, and IT is a critical competence for all enterprises.
 Datagrok enables IT to become a partner with business by solving business problems, at the same time 
 providing the following:
 * Enterprise-grade [security](/help/govern/security)
@@ -95,7 +94,8 @@ providing the following:
 * Data governance, data lineage, and [audit](/help/govern/audit)
 * Lightweight, managed deployment of scientific methods and predictive models
 * [Usage analysis](/help/govern/usage-analysis)`;
-var introDS = `
+
+const introDS = `
 Datagrok lets you easily connect to data, visualize and explore it, build predictive models, 
 and deploy your work to production at a speed previously unheard of. Unlike other platforms,
 "deployment" in Datagrok is not simply exposing an endpoint API. \n 
@@ -107,14 +107,16 @@ and end up training a predictive model and producing a rich application using it
 * Enrich db query results with the augmented data, and give them to the end users 
 * Unmatched capabilities for [exploratory data analysis](/help/explore/exploratory-data-analysis)
 * Deploy [R, Python, Julia](/help/compute/scripting) scripts into production''';`
-var introChem = `''Datagrok provides [first-class support for small molecules](/help/domains/chem/cheminformatics).
+
+const introChem = `''Datagrok provides [first-class support for small molecules](/help/domains/chem/cheminformatics).
 * [Chemically aware viewers](/help/domains/chem/chemically-aware-viewers)
 * Rich set of [molecular descriptors](/help/domains/chem/descriptors)
 * Support for RDKit or OpenChemLib  
 * [Sketch molecules](/help/domains/chem/sketcher) 
 * [R-group analysis](/help/domains/chem/r-group-analysis)
 * Chemical [Similarity](/help/domains/chem/similarity-search) and [diversity](/help/domains/chem/diversity-search) analyses`;
-var introDev = `
+
+const introDev = `
 Think about Datagrok as an OS for data. Just like the way Linux enables application developers to be productive
 by handling low-level stuff and introducing high-level concepts like processes, pipes, and files, Datagrok
 does the same for data. By leveraging our proprietary high-performance [data engine](/help/develop/performance), we provide
@@ -126,7 +128,8 @@ This lightweight process will take you a long way; some applications that take m
 now be [done in a matter of days](https://www.youtube.com/watch?v=tVwpRB8fikQ). \n
 For deeper integration and customization, use our [JS API](/help/develop/develop) to take full control
 of the platform.`;
-var introLeaders = `Connect business areas with IT, data science, and application developers.
+
+const introLeaders = `Connect business areas with IT, data science, and application developers.
 * Take control of your data and analytics landscape by connecting with existing systems
 * Break organizational barriers by combining data and algorithms from disparate data silos and functional areas
 * Start small, and organically grow your ecosystem